requirements.txt
PySide6==6.9.2
requests>=2.31.0
python-dotenv>=1.0.0
